task creation with temp dumb id generation; better understanding of thymeleaf and web dev

+++ src/main/kotlin/com/kasymzhan/springMongo/repository/TaskRepository.kt
@@ -3,4 +3,4 @@ package com.kasymzhan.springMongo.repository
 import com.kasymzhan.springMongo.model.Task
 import org.springframework.data.mongodb.repository.MongoRepository
 
-interface TaskRepository : MongoRepository<Task, Int> {}
+interface TaskRepository : MongoRepository<Task, Long> {}
